+++ package.json
@@ -57,7 +57,8 @@
   "dependencies": {
     "@curveball/http-errors": "^0.3.0",
     "accepts": "^1.3.5",
-    "raw-body": "^2.4.0"
+    "raw-body": "^2.4.0",
+    "structured-headers": "^0.2.0"
   },
   "engines": {
     "node": ">=9.4"
