+++ src/memory-request.ts
@@ -1,10 +1,7 @@
-import accepts from 'accepts';
-import url from 'url';
-import { is } from './header-helpers';
 import { Headers, HeadersInterface, HeadersObject } from './headers';
 import Request from './request';
 
-export class MemoryRequest<T> implements Request<T> {
+export class MemoryRequest<T> extends Request<T> {
 
   /**
    * List of HTTP Headers
@@ -19,6 +16,7 @@ export class MemoryRequest<T> implements Request<T> {
 
   constructor(method: string, requestTarget: string, headers?: HeadersInterface | HeadersObject, body: any = null) {
 
+    super();
     this.method = method;
     this.requestTarget = requestTarget;
     if (headers && (<HeadersInterface> headers).get !== undefined) {
@@ -32,53 +30,6 @@ export class MemoryRequest<T> implements Request<T> {
     this.body = null;
 
   }
-
-  /**
-   * path-part of the request.
-   *
-   * For example /hello/world
-   */
-  get path(): string {
-
-    return url.parse(this.requestTarget).pathname!;
-
-  }
-
-  /**
-   * Sets the path
-   */
-  set path(value: string) {
-
-    this.requestTarget = value;
-
-  }
-
-  method: string;
-
-  /**
-   * The request target.
-   *
-   * This contains the literal value after the HTTP method in the request.
-   * So for:
-   *
-   * GET /foo HTTP/1.1
-   *
-   * This would contain '/foo'. In many cases this is the same as the 'path'
-   * property, but there's 3 other forms in the HTTP specificatio. Here's the
-   * different formats:
-   *
-   * * origin-form    - This is the most common. Example: /foo.
-   * * absolute-form  - Clients might sent an entire path. Also used by proxies.
-   *                    Example: https://example.org/foo
-   * * authority-form - Used by the CONNECT method. Example: example.org:1212
-   * * asterisk-form  - Used by the OPTIONS method. Example: *
-   *
-   * In most cases users will want to use the 'path' property instead. only use
-   * this if you know what you're doing.
-   *
-   * @see {@link https://tools.ietf.org/html/rfc7230#section-5.3}
-   */
-  requestTarget: string;
 
   /**
    * Internal representation of body.
@@ -128,73 +79,6 @@ export class MemoryRequest<T> implements Request<T> {
         return this.originalBody;
       }
     }
-
-  }
-
-  /**
-   * This object contains parsed query parameters.
-   */
-  get query(): { [s: string]: string } {
-
-    return <any> url.parse(this.requestTarget, true).query;
-
-  }
-
-  /**
-   * Returns the value of the Content-Type header, with any additional
-   * parameters such as charset= removed.
-   *
-   * If there was no Content-Type header, an empty string will be returned.
-   */
-  get type(): string {
-
-    const type = this.headers.get('content-type');
-    if (!type) { return ''; }
-    return type.split(';')[0];
-
-  }
-
-  /**
-   * accepts is used for negotation the Content-Type with a client.
-   *
-   * You can pass a content-type, or an array of content-types.
-   * The Content-Types you provide are a list of types your application
-   * supports.
-   *
-   * This function will then return the best possible type based on the Accept
-   * header.
-   *
-   * If no compatible types are found, this function returns null.
-   */
-  accepts(...types: string[]): null | string {
-
-    const dummyRequest: any = {
-      headers: {}
-    };
-    const acceptHeader = this.headers.get('accept');
-    if (acceptHeader) {
-      dummyRequest.headers.accept = acceptHeader;
-    }
-    const result = <string|false> accepts(dummyRequest).type(types);
-    return result === false ? null : result;
-
-  }
-
-  /**
-   * This method will return true or false if a Request or Response has a
-   * Content-Type header that matches the argument.
-   *
-   * For example, if the Content-Type header has the value: application/hal+json,
-   * then the arguments will all return true:
-   *
-   * * application/hal+json
-   * * application/json
-   * * hal+json
-   * * json
-   */
-  is(type: string): boolean {
-
-    return is(this, type);
 
   }
